Working SSH-tunnel

+++ dask_cloudprovider/providers/azure/azureml.py
@@ -3,12 +3,13 @@ import logging
 from azureml.core import Experiment
 from azureml.train.estimator import Estimator
 from azureml.core.runconfig import MpiConfiguration
-import time, os, socket, sys
+import time, os, socket, sys, subprocess
 
 from distributed.deploy.cluster import Cluster
 from distributed.core import rpc
 
 import dask
+import threading
 
 from distributed.utils import (
     LoopRunner,
@@ -78,7 +79,22 @@ class AzureMLCluster(Cluster):
         Port to map the scheduler port to via SSH-tunnel if machine not on the same VNET. 
 
         Defaults to ``9002``.
+        
+    admin_username: str (optional)
+        Username of the admin account for the AzureML Compute.
+        Required for runs that are not on the same VNET. Defaults to empty string. 
+        Throws Exception if machine not on the same VNET.
 
+        Defaults to ``""``.
+        
+    admin_ssh_key: str (optional)
+        Location of the SSH secret key used when creating the AzureML Compute.
+        The key should be passordless if run from a Jupyter notebook.
+        Required for runs that are not on the same VNET. Defaults to empty string. 
+        Throws Exception if machine not on the same VNET.
+
+        Defaults to ``""``.
+        
     datastores: List[str] (optional)
         List of Azure ML Datastores to be mounted on the headnode - 
         see https://aka.ms/azureml/data and https://aka.ms/azureml/datastores. 
@@ -142,6 +158,8 @@ class AzureMLCluster(Cluster):
         , jupyter_port=None             # port to forward the Jupyter process to
         , dashboard_port=None           # port to forward Dask dashboard to
         , scheduler_port=None           # port to map the scheduler port to for 'local' runs
+        , admin_username=None           # username to log in to the AzureML Training Cluster for 'local' runs
+        , admin_ssh_key=None            # path to private SSH key used to log in to the AzureML Training Cluster for 'local' runs
         , datastores=None               # datastores specs
         , code_store=None               # name of the code store if specified
         , asynchronous=True             # flag to run jobs in an asynchronous way
@@ -167,6 +185,8 @@ class AzureMLCluster(Cluster):
         self.jupyter_port = jupyter_port
         self.dashboard_port = dashboard_port
         self.scheduler_port = scheduler_port
+        self.admin_username = admin_username
+        self.admin_ssh_key = admin_ssh_key
         self.scheduler_ip_port = None   ### INIT FOR HOLDING THE ADDRESS FOR THE SCHEDULER
         
         ### DATASTORES
@@ -217,6 +237,12 @@ class AzureMLCluster(Cluster):
             
         if self.scheduler_port is None:
             self.scheduler_port = self.config.get("scheduler_port")
+            
+        if self.admin_username is None:
+            self.admin_username = self.config.get("admin_username")
+            
+        if self.admin_ssh_key is None:
+            self.admin_ssh_key = self.config.get("admin_ssh_key")
 
         if self.datastores is None:
             self.datastores = self.config.get("datastores")
@@ -273,11 +299,20 @@ class AzureMLCluster(Cluster):
             ip, port = self.scheduler_ip_port.split(':')
             socket.create_connection((ip, port), 5)
             self.same_vnet = True
-            print('Same VNET')
+            self.__print_message('On the same VNET')
         except socket.timeout:
-            print('Not on the same vnet')
+            self.__print_message('Not on the same VNET')
             self.same_vnet = False
-
+            
+    def __prepare_rpc_connection_to_headnode(self):
+        if not self.same_vnet:
+            if (self.admin_username == "" or self.admin_ssh_key == ""):
+                raise Exception("Your machine is not at the same VNET as the cluster. You need to set admin_username and admin_ssh_key. Check documentation.")
+            else:
+                return f"{socket.gethostname()}:{self.scheduler_port}"
+        else:
+            return self.run.get_metrics()["scheduler"]
+        
     async def create_cluster(self):
         # set up environment
         self.__print_message('Setting up cluster')
@@ -304,8 +339,8 @@ class AzureMLCluster(Cluster):
              run.get_status() != 'Canceled'
              and run.get_status() != 'Failed'
              and 'scheduler' not in run.get_metrics()):
-             print('.', end="")
-             time.sleep(5)
+                 print('.', end="")
+                 time.sleep(5)
             
         if run.get_status() == 'Canceled' or run.get_status() == 'Failed':
             raise Exception('Failed to start the AzureML cluster.')
@@ -318,16 +353,27 @@ class AzureMLCluster(Cluster):
         self.__print_message(f'Scheduler: {run.get_metrics()["scheduler"]}')
         self.run = run
         
+        ### CHECK IF ON THE SAME VNET
+        while(self.same_vnet is None):
+            try:
+                await self.sync(self.__check_if_scheduler_ip_reachable)
+            except:
+                time.sleep(1)
+                
         ### REQUIRED BY dask.distributed.deploy.cluster.Cluster
-        self.scheduler_comm = rpc(run.get_metrics()["scheduler"])
+        _scheduler = self.__prepare_rpc_connection_to_headnode()
+        self.scheduler_comm = rpc(_scheduler)
         # asyncio.ensure_future(super()._start())
+        await self.sync(self.__setup_port_forwarding)
         await self.sync(super()._start)
-        await self.sync(self.__check_if_scheduler_ip_reachable)
+        await self.sync(self.__update_links)
         
-        self.__print_message(f'Scaling to {self.initial_node_count} workers')
-        if self.initial_node_count > 1:
-            self.scale(self.initial_node_count)   # LOGIC TO KEEP PROPER TRACK OF WORKERS IN `scale`
-        self.__print_message(f'Scaling is done')
+        self.__print_message('Connections established')
+                
+#         self.__print_message(f'Scaling to {self.initial_node_count} workers')
+#         if self.initial_node_count > 1:
+#             self.scale(self.initial_node_count)   # LOGIC TO KEEP PROPER TRACK OF WORKERS IN `scale`
+#         self.__print_message(f'Scaling is done')
         
     def connect_cluster(self):
         if not self.run:
@@ -338,16 +384,48 @@ class AzureMLCluster(Cluster):
             print(f'\nSetting up port forwarding...')
             self.port_forwarding_ComputeVM()
             print(f'Cluster is ready to use.')
+            
+    async def __update_links(self):
+        if self.same_vnet:
+            self.scheduler_info['dashboard_url'] = f'https://{socket.gethostname()}-{self.dashboard_port}.{self.workspace.get_details()["location"]}.instances.azureml.net/status'
+            self.scheduler_info['jupyter_url'] = (
+                f'https://{socket.gethostname()}-{self.jupyter_port}.{self.workspace.get_details()["location"]}.instances.azureml.net/lab?token={self.run.get_metrics()["token"]}'
+            )
+        else:
+            self.scheduler_info['dashboard_url'] = f'http://{socket.gethostname()}:{self.dashboard_port}'
+            self.scheduler_info['jupyter_url']   = f'http://{socket.gethostname()}:{self.jupyter_port}/?token={self.run.get_metrics()["token"]}'
+                
+#         print(self.scheduler_info['jupyter_url'])
+#         print(self.scheduler_info['dashboard_url'])
 
-    def port_forwarding_ComputeVM(self):
-        os.system(f'killall socat') # kill all socat processes - cleans up previous port forward setups
-        os.system(f'setsid socat tcp-listen:{self.dashboard_port},reuseaddr,fork tcp:{self.run.get_metrics()["dashboard"]} &')
-        os.system(f'setsid socat tcp-listen:{self.jupyter_port},reuseaddr,fork tcp:{self.run.get_metrics()["jupyter"]} &')
-        
-        self.scheduler_info['dashboard_url'] = f'https://{socket.gethostname()}-{self.dashboard_port}.{self.workspace.get_details()["location"]}.instances.azureml.net/status'
-        self.scheduler_info['jupyter_url'] = (
-            f'https://{socket.gethostname()}-{self.jupyter_port}.{self.workspace.get_details()["location"]}.instances.azureml.net/lab?token={self.run.get_metrics()["token"]}'
-        )
+    async def __setup_port_forwarding(self):
+        if self.same_vnet:
+            os.system(f'killall socat') # kill all socat processes - cleans up previous port forward setups
+            os.system(f'setsid socat tcp-listen:{self.dashboard_port},reuseaddr,fork tcp:{self.run.get_metrics()["dashboard"]} &')
+            os.system(f'setsid socat tcp-listen:{self.jupyter_port},reuseaddr,fork tcp:{self.run.get_metrics()["jupyter"]} &')
+        else:
+            scheduler_ip = self.run.get_metrics()["scheduler"].split(':')[0]
+            scheduler_public_ip   = self.compute_target.list_nodes()[0]['publicIpAddress']
+            scheduler_public_port = self.compute_target.list_nodes()[0]['port']
+            
+            cmd = ("ssh -vvv -o StrictHostKeyChecking=no -N" + \
+               f" -i {self.admin_ssh_key}"
+               f" -L 0.0.0.0:{self.jupyter_port}:{scheduler_ip}:8888" + \
+               f" -L 0.0.0.0:{self.dashboard_port}:{scheduler_ip}:8787" + \
+               f" -L 0.0.0.0:{self.scheduler_port}:{scheduler_ip}:8786" + \
+               f" {self.admin_username}@{scheduler_public_ip} -p {scheduler_public_port}"
+            )
+            
+            portforward_log = open("portforward_out_log.txt", 'w')
+            portforward_proc = (
+                subprocess
+                .Popen(
+                    cmd.split()
+                    , universal_newlines=True
+                    , stdout=subprocess.PIPE
+                    , stderr=subprocess.STDOUT
+                )
+            ) 
         
     @property
     def dashboard_link(self):
